ovs-numa: Non-consecutively numa nodes and cores
This change removes the assumption that numa nodes and cores are numbered
consecutively in linux.  This change is required to support some Power
systems.

An additional check has been added to verify that cores are online,
offline cores result in non-consecutively numbered cores.

I manually verified that all the exported functions defined in
ovs-numa.h function correctly with these changes on Power and amd64
systems.

Tests for dpif-netdev and pmd that define numa systems using the
--dummy-numa option have been modified to use non-consecutive numbered
nodes.

Changes for v2:
0-day Robot suggested changes.

Signed-off-by: David Wilder <dwilder@us.ibm.com>